added error handling and validation for page size and number

+++ src/Services/artwork/ArtworkService.cs
@@ -1,25 +1,23 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using AutoMapper;
 using Backend_Teamwork.src.Entities;
 using Backend_Teamwork.src.Repository;
-using Backend_Teamwork.src.Services.artwork;
 using Backend_Teamwork.src.Utils;
-using Microsoft.EntityFrameworkCore.Update.Internal;
 using static Backend_Teamwork.src.DTO.ArtworkDTO;
+using static Backend_Teamwork.src.DTO.UserDTO;
+using static Backend_Teamwork.src.Entities.User;
 
 namespace Backend_Teamwork.src.Services.artwork
 {
     public class ArtworkService : IArtworkService
     {
         private readonly ArtworkRepository _artworkRepo;
+        private readonly UserRepository _userRepo;
         private readonly IMapper _mapper;
 
-        public ArtworkService(ArtworkRepository artworkRepo, IMapper mapper)
+        public ArtworkService(ArtworkRepository artworkRepo, UserRepository userRepo, IMapper mapper)
         {
             _artworkRepo = artworkRepo;
+            _userRepo = userRepo;
             _mapper = mapper;
         }
 
@@ -33,6 +31,17 @@ namespace Backend_Teamwork.src.Services.artwork
 
         public async Task<List<ArtworkReadDto>> GetAllAsync(PaginationOptions paginationOptions)
         {
+            // Validate pagination options
+            if (paginationOptions.PageSize <= 0)
+            {
+                throw CustomException.BadRequest("PageSize should be greater than 0.");
+            }
+
+            if (paginationOptions.PageNumber <= 0)
+            {
+                throw CustomException.BadRequest("Offset should be greater than 0.");
+            }
+
             var artworkList = await _artworkRepo.GetAllAsync(paginationOptions);
             return _mapper.Map<List<Artwork>, List<ArtworkReadDto>>(artworkList);
         }
@@ -40,12 +49,22 @@ namespace Backend_Teamwork.src.Services.artwork
         public async Task<ArtworkReadDto> GetByIdAsync(Guid id)
         {
             var artwork = await _artworkRepo.GetByIdAsync(id);
-            //  TO DO: handle error
+            if (artwork == null) {
+                throw CustomException.NotFound($"Artwork with id: {id} not found");
+            }
             return _mapper.Map<Artwork, ArtworkReadDto>(artwork);
         }
 
         public async Task<List<ArtworkReadDto>> GetByArtistIdAsync(Guid id){
-            var artworks = await _artworkRepo.GetByArtistIdAsync(id);
+            // check if user exist
+            var user = await _userRepo.GetByIdAsync(id) ?? throw CustomException.NotFound($"User with id: {id} not found");
+            // check the role of user
+            if (user.Role.ToString() != UserRole.Artist.ToString())
+            {
+                throw CustomException.BadRequest($"User with id: {id} is not an Artist");
+            }
+            // check if user(artist) has artwork
+            var artworks = await _artworkRepo.GetByArtistIdAsync(id) ?? throw CustomException.NotFound($"Artist with id: {id} has no artworks");
             var artworkList = _mapper.Map<List<Artwork>, List<ArtworkReadDto>>(artworks);
             return artworkList;
         }
@@ -53,6 +72,10 @@ namespace Backend_Teamwork.src.Services.artwork
         public async Task<bool> DeleteOneAsync(Guid id)
         {
             var foundArtwork = await _artworkRepo.GetByIdAsync(id);
+            if (foundArtwork == null)
+            {
+                throw CustomException.NotFound($"Artwork with id: {id} not found");
+            }
             bool isDeleted = await _artworkRepo.DeleteOneAsync(foundArtwork);
 
             return isDeleted;
@@ -63,10 +86,9 @@ namespace Backend_Teamwork.src.Services.artwork
             var foundArtwork = await _artworkRepo.GetByIdAsync(id);
             if (foundArtwork == null)
             {
-                return false;
+                throw CustomException.NotFound($"Artwork with id: {id} not found");
             }
 
-            // keep old one - update a part of it
             _mapper.Map(updateDto, foundArtwork);
             return await _artworkRepo.UpdateOneAsync(foundArtwork);
         }
